g104 - Rename routing
Description
Renamed and change file and template names of
home to landing
memo.feature to memo.home
memo.home to gallery
to make sense
