--- imports/ui/featured/Featured.html
@@ -1,77 +1,0 @@
-<template name="Featured">
-	<div class="main-container">
-		{{> PageTitle}}
-		<div class="main-content">
-		{{#if Template.subscriptionsReady}}
-		<div class="featured-collection">
-			<ul class="collection">
-				<li class="collection-item">
-					<span class="title">{{i18n 'feature.favorites'}}</span>
-					<div class="row">
-						{{#each favoriteMemos}}
-							{{> Memo}}
-						{{else}}
-							<span class="grey-text">{{i18n 'feature.noFavorite'}}</span>
-						{{/each}}
-						{{#if $gt state.favoriteCount 8}}
-							<div class="col s12 center">
-								{{#if favoriteHasMoreContent}}
-									<button class="btn white black-text show-more-favorite">
-										{{i18n 'feature.more'}}
-									</button>
-								{{else}}
-									<button class="btn white black-text hide-favorite">
-										{{i18n 'feature.hide'}}
-									</button>
-								{{/if}}
-							</div>
-						{{/if}}
-					</div>
-				</li>
-				<li class="collection-item">
-					<span class="title">{{i18n 'feature.recentlyViewed'}}</span>
-					<div class="row">
-						{{#each recentMemos}}
-							{{> Memo}}
-						{{else}}
-							<span class="grey-text">{{i18n 'feature.noRecent'}}</span>
-						{{/each}}
-						{{#if $gt state.recentCount 8}}
-							<div class="col s12 center">
-								{{#if recentHasMoreContent}}
-									<button class="btn white black-text show-more-recent">
-										{{i18n 'feature.more'}}
-									</button>
-								{{else}}
-									<button class="btn white black-text hide-recent">
-										{{i18n 'feature.hide'}}
-									</button>
-								{{/if}}
-							</div>
-						{{/if}}
-					</div>
-				</li>
-				<li class="collection-item">
-					<span class="title">{{i18n 'feature.forYou'}}
-						<span class="chip" style="background-color:{{recommendLabel.color}};">{{recommendLabel.name}}</span>
-					</span>
-					<div class="row">
-						{{#each recommendMemos}}
-							{{> Memo}}
-						{{else}}
-							<span class="grey-text">{{i18n 'feature.noRecommend'}}</span>
-						{{/each}}
-					</div>
-				</li>
-			</ul>
-		</div>
-		{{else}}
-			<div class="loading-wrapper valign-wrapper">
-				<div class="center valign">
-					{{> Loading size="big"}}
-				</div>
-			</div>
-		{{/if}}
-		</div>
-	</div>
-</template>
