+++ imports/ui/gallery/Gallery.html
@@ -1,4 +1,4 @@
-<Template name="Memos">
+<Template name="Gallery">
     <div class="main-container">
     {{> PageTitle}}
         <div class="main-content">
