+++ imports/ui/home/Home.html
@@ -1,40 +1,77 @@
 <template name="Home">
-		<section class="landing-page-section center">
-			<h1 class="header">DevMemo</h1>
-			<h4 class="light">{{i18n 'home.subtitle'}}</h4>
-			<a href="{{pathFor route='atSignIn'}}" class="waves-effect btn btn-lg blue darken-3">{{i18n 'home.gettingStarted'}}</a>
-		</section>
-		<section class="landing-page-section light-blue lighten-5">
-			<div class="container landing-container">
-				<div class="row">
-					<div class="col m4">
-						<div class="icon-block">
-							<h2 class="center blue-text">
-								<i class="fa fa-user"></i>
-							</h2>
-							<h5 class="center">{{i18n 'home.easyToUse.title'}}</h5>
-							<p class="light">{{i18n 'home.easyToUse.description'}}</p>
-						</div>
+	<div class="main-container">
+		{{> PageTitle}}
+		<div class="main-content">
+		{{#if Template.subscriptionsReady}}
+		<div class="featured-collection">
+			<ul class="collection">
+				<li class="collection-item">
+					<span class="title">{{i18n 'feature.favorites'}}</span>
+					<div class="row">
+						{{#each favoriteMemos}}
+							{{> Memo}}
+						{{else}}
+							<span class="grey-text">{{i18n 'feature.noFavorite'}}</span>
+						{{/each}}
+						{{#if $gt state.favoriteCount 8}}
+							<div class="col s12 center">
+								{{#if favoriteHasMoreContent}}
+									<button class="btn white black-text show-more-favorite">
+										{{i18n 'feature.more'}}
+									</button>
+								{{else}}
+									<button class="btn white black-text hide-favorite">
+										{{i18n 'feature.hide'}}
+									</button>
+								{{/if}}
+							</div>
+						{{/if}}
 					</div>
-					<div class="col m4">
-						<div class="icon-block">
-							<h2 class="center blue-text">
-								<i class="fa fa-picture-o"></i>
-							</h2>
-							<h5 class="center">{{i18n 'home.wellDesigned.title'}}</h5>
-							<p class="light">{{i18n 'home.wellDesigned.description'}}</p>
-						</div>
+				</li>
+				<li class="collection-item">
+					<span class="title">{{i18n 'feature.recentlyViewed'}}</span>
+					<div class="row">
+						{{#each recentMemos}}
+							{{> Memo}}
+						{{else}}
+							<span class="grey-text">{{i18n 'feature.noRecent'}}</span>
+						{{/each}}
+						{{#if $gt state.recentCount 8}}
+							<div class="col s12 center">
+								{{#if recentHasMoreContent}}
+									<button class="btn white black-text show-more-recent">
+										{{i18n 'feature.more'}}
+									</button>
+								{{else}}
+									<button class="btn white black-text hide-recent">
+										{{i18n 'feature.hide'}}
+									</button>
+								{{/if}}
+							</div>
+						{{/if}}
 					</div>
-					<div class="col m4">
-						<div class="icon-block">
-							<h2 class="center blue-text">
-								<i class="fa fa-refresh"></i>
-							</h2>
-							<h5 class="center">{{i18n 'home.realTime.title'}}</h5>
-							<p class="light">{{i18n 'home.realTime.description'}}</p>
-						</div>
+				</li>
+				<li class="collection-item">
+					<span class="title">{{i18n 'feature.forYou'}}
+						<span class="chip" style="background-color:{{recommendLabel.color}};">{{recommendLabel.name}}</span>
+					</span>
+					<div class="row">
+						{{#each recommendMemos}}
+							{{> Memo}}
+						{{else}}
+							<span class="grey-text">{{i18n 'feature.noRecommend'}}</span>
+						{{/each}}
 					</div>
+				</li>
+			</ul>
+		</div>
+		{{else}}
+			<div class="loading-wrapper valign-wrapper">
+				<div class="center valign">
+					{{> Loading size="big"}}
 				</div>
 			</div>
-		</section>
+		{{/if}}
+		</div>
+	</div>
 </template>
