+++ imports/ui/landing/Landing.html
@@ -1,0 +1,40 @@
+<template name="Landing">
+		<section class="landing-page-section center">
+			<h1 class="header">DevMemo</h1>
+			<h4 class="light">{{i18n 'home.subtitle'}}</h4>
+			<a href="{{pathFor route='atSignIn'}}" class="waves-effect btn btn-lg blue darken-3">{{i18n 'home.gettingStarted'}}</a>
+		</section>
+		<section class="landing-page-section light-blue lighten-5">
+			<div class="container landing-container">
+				<div class="row">
+					<div class="col m4">
+						<div class="icon-block">
+							<h2 class="center blue-text">
+								<i class="fa fa-user"></i>
+							</h2>
+							<h5 class="center">{{i18n 'home.easyToUse.title'}}</h5>
+							<p class="light">{{i18n 'home.easyToUse.description'}}</p>
+						</div>
+					</div>
+					<div class="col m4">
+						<div class="icon-block">
+							<h2 class="center blue-text">
+								<i class="fa fa-picture-o"></i>
+							</h2>
+							<h5 class="center">{{i18n 'home.wellDesigned.title'}}</h5>
+							<p class="light">{{i18n 'home.wellDesigned.description'}}</p>
+						</div>
+					</div>
+					<div class="col m4">
+						<div class="icon-block">
+							<h2 class="center blue-text">
+								<i class="fa fa-refresh"></i>
+							</h2>
+							<h5 class="center">{{i18n 'home.realTime.title'}}</h5>
+							<p class="light">{{i18n 'home.realTime.description'}}</p>
+						</div>
+					</div>
+				</div>
+			</div>
+		</section>
+</template>
